Add support for @provideAlreadyProvided
When this JSDoc annotation is used with a `goog.provide` call like:

```
goog.provide('a.b');
/** @provideAlreadyProvided */ goog.provide('a.b.c');
a.b = something;
```

The `goog.provide` rewriting pass (ProcessClosureProvidesAndRequires) will avoid creating the definition for the namespace.  Without this the pass would create a declaration (`a.b.c = {}` in the example above).

This solution was chosen over alternatives, as generally speaking, `goog.provide` is deprecated in favor of `goog.module` so more structural changes (such as never creating implicit namespaces) are not worthwhile due to the migration costs.

PiperOrigin-RevId: 435677611