Loosen the constraints on `goog.module.get` so that only assignments to global variables are banned.
The constraint on where `goog.module.get` was called was to discourage accidental treatment of `goog.provide` or bare scripts as having module scopes and polluting the global scope with imports and causing problems in far flung files.

However, we want to allow reexporting of `goog.module` namespaces in goog.provide files to simplify migrations to `goog.module`.  So here we explicitly allow alias such as:

```
goog.provide('a.b.c');
a.b.c = goog.module.get('c')
```

or

```
goog.provide('a.b.c');
a.b.c = goog.module.get('other').c;
```

While we are here cleanup the checks for `goog.module.get` so that they are only reported from CheckClosureImports.

PiperOrigin-RevId: 436496743