Add another case to OptimizeParameters to eliminate trailing explicit `undefined` values when calling functions that don't reference `arguments`
In hand written code this pattern is unusual but can easily be created by inlining functions that pass optional parameters through.

for example,

```
function foo(opt) {
...
}

function caller(opt) { foo(opt); }
foo('a');
caller();
```

Due to inlining, that might get optimized as:

```
function foo(opt) {
...
}
foo('a');
foo(undefined);
```

even though that `undefined` is not necessary.

With this change, `OptimizeParameters` will remove 'trailing undefined' values as long as the callee function does not reference `arguments`, since in that scenario passing/not-passing the parameter should not be observable.

PiperOrigin-RevId: 437849357