`InlineVariables`: Inline aliases in a single pass
Consider this code:

```javascript
const A = 1;

function foo() {
  const a = A;
  const arrow = () => {
    const innerA = a;
    console.log(innerA);
  }
}
```

There is a chain of aliases here.

`innerA` is an alias of `a`.

`a` is an alias of `A`.

To keep memory requirements in check `InlineVariables` needs
to consider variables starting with the innermost scope and working outwards.

Prior to this change, 3 executions could be required to inline
all of these variables, because inlining of an inner variable
alias caused the outer, aliased variable to be marked "stale"
and ineligible for inlining for the rest of this run.

Actually, that's not quite true, since there was special-purpose
code that tried to recognize aliases and handle them a bit
better than that.

Still, this was inefficient and prioritized the inner variables over the
outer ones, which is less beneficial than the reverse.

Starting with this change, `InlineVariables` will delay its decision
for `innerA` until it has decided on `a`.

In turn the decision for `a` is delayed until `A` is handled.

`A` gets inlined, yielding `const a = 1`.

`a` is reconsidered & inlined, yielding `innerA = 1`.

Finally, `innerA` is inlined, yielding `console.log(1)`

All of this happens within a single execution of the pass.

PiperOrigin-RevId: 439931473