Improve behavior of printing chunks after each pass
Previously when a list of chunk name regular expressions were given for printing after each pass,
but none of the actual chunks matched any of those patterns, the compiler exited with an exception
that listed the patterns.

This is unhelpful in 2 ways.

1. It's easy to see what the patterns were by looking at the command line or
   compiler hack you probably just wrote.  It would be much more useful to
   print the list of actual chunk names, since apparently you were wrong in
   what you expected them to be.
2. We often need to launch a build that may invoke the compiler in many different ways,
   only one of which will involve the chunks we care about. Throwing an exception
   kills the entire build, which is not helpful.

So, now when none of the chunks are matched the compiler prints the list of
unmatched chunks where it would normally print their contents and does not fail.

PiperOrigin-RevId: 442079726