InlineVariables: Remove dead string inlining logic
Delete the `inlineAllStrings` field from `InlineVariables` and the
code paths that were enabled when it was `false`.

AFAICT it has been hard-coded to `true` everywhere except for some unit tests
since at least 2009.

We do want to inline all strings, because inlining them often allows us to
pre-compute conditions and so eliminate dead code paths.

`AliasStrings` exists to undo this inlining when that is desirable.

PiperOrigin-RevId: 442562157