Delete dead CompilerOptions field
`CompilerOptions.externExports` was only ever set to 'true' in tests
in order to enable something not really related to creating externs.
The `externExportsPath` field is the real indicator of whether we
are generating externs for exports.

Also, normalize access to the externExportsPath field to use
a getter and a setter instead of the inconsistent mix we had before.

PiperOrigin-RevId: 443748754