Drop source code data from ijs files after parsing
It is my understanding that the source code caches inside of `SourceFile` are intended for cases like error reporting and snippet generation.  In general, I would guess that this is a suspicious optimization (error formatting should be 'rare', so making it faster probably isn't important).  But in the case of `@typeSummary` files error reporting is nearly nonexistant. So caching the code is almost always a waste of ram.

PiperOrigin-RevId: 444938446