Improve quadratic performance edge case in JSType->Color serialization
The problem: structural type hash codes are only based on their set of properties. This led to >4k structural JSTypes being hashed to the same bucket in the map in InvalidatingTypes for a very large Angular app.

The fix: avoid adding structural types to the map of invalidating types, as we already know they are invalidating b/c they're structural.

Note: in theory we could fix structural type hash codes to depend on property types as well as names, but I think that would be more expensive & also cause potential issues with cyclic type references.
PiperOrigin-RevId: 448063787