Fix some bugs in constant folding for ??
In PeepholeRemoveDeadCode: fix incorrect replacement of the condition in `a ?? <expr>` with the <primitive> value coerced to boolean.

In PeepholeFoldConstants: fix case where if the left side was not nullish and had side effects, the expression was incorrectly folded to evaluate to the right side

Fixes https://github.com/google/closure-compiler/issues/3943.

PiperOrigin-RevId: 449513603