Correct the exit status for partial compilations
A recent change to ensure partial compilations (stage 1 or stage 2) do not
generate extra output files accidentally caused them to exit with a zero status
even when the compiler reported errors in the code.

This change fixes that problem.
The compiler again exits with a 0 status only when no errors are reported.
The exit status contains the number of errors reported up to a maximum of 127.

PiperOrigin-RevId: 457579936