Stop serializing `Compiler#changeStamp`
Since we aren't serializing the corresponding properties on the AST nodes,
there's no benefit in serializing `changeStamp` between stages.
We can just re-initialize it to 1 when restoring state.

Also, it turns out that for some unknown reason the value of `changeStamp`
isn't deterministic, which made the `.save` file contents non-deterministic,
which caused problems for build systems using closure-compiler.

PiperOrigin-RevId: 460527162