Simplify deserialization of properties logic
Do a single loop and switch over the bits of a long to deserialze properties.

This removes some awkwardness in logic that handles the one bit field property
we currently serialize.

It doesn't matter much for just that one 2-bit field, but I'm about to add a
4-bit field, which would be much more awkward if I followed the older pattern.

The only behavior change in this CL is that the CONSTANT_VAR_FLAGS property
ends up at a different position in the property list.

PiperOrigin-RevId: 471649796