Make the `setOutputFeatureSet` API package private
Since the `setOuputFeatureSet` API allow setting output modes that are not
properly handled, it is being removed from the public Java API, in favor
of better APIs like `setBrowserFeaturesetYear` and `setLangauageOut` which
only allow setting options that are properly handled.

This also migrates the last out-of-package usage in the transpile package to
use a @RestrictedApi instead to make it clearer that it is not supported.

PiperOrigin-RevId: 471667019