Factor logic out of JsMessage#Builder.build()
JsMessage#Builder.build() contained quite a bit of logic to infer
the intentions of the code using the builder based on which fields
were set. This obscured the decisions that were really being made
by the client code, making it both harder to understand and harder to change.

This refactoring should not have any runtime behavior impact.
It is needed to prepare for changes to the logic that was previously
hidden inside JsMeassge#Builder.build().

PiperOrigin-RevId: 471876170