Make the choice of placeholder name format explicit
Split `getPlaceholderName()` method into `getJsPlaceholderName()`
and `getCanonicalPlaceholderName()` to make it clear which format is needed
at the usage site.

Also split `JsMessage.Builder.appendPlaceholderName()` into "Js" and
"Canonical" forms to make it explicitly clear what format the name used to
create the placeholder is in.

This is just a refactoring and should have no meaningful change on execution.

PiperOrigin-RevId: 474195237