Remove unnecessary default branches in switch statements
The error prone tool that we build with already gives an error if a switch
statement over an enum isn't exhaustive. This means that removing the default
cases here will upgrade the runtime error that we would have gotten if
a new item was added to one of these enums into a build breakage instead.

PiperOrigin-RevId: 474691845