Stop automatically converting ICU message placeholders to strings
This change moves the logic for deciding how to handle placeholders in
ICU-selector formatted messages from the bundle-reading code to the
ReplaceMessages pass.

Prior to this change the code for reading translated message bundles would
recognize ICU-selector formatted messages and automatically convert all of the
placeholder elements in them into string representations of the
placeholders. (e.g. A placeholder named FOO becomes just a string '{FOO}')

This was because the compiler doesn't really understand ICU messages. When they
are passed to `goog.getMsg()` the compiler sees them as simple strings with no
placeholders. This "hide the placeholders" behavior existed to handle the case
where the message bundle was generated by something other than
closure-compiler, so it contained placeholders, which the compiler wouldn't be
able to match with the JS code version of the same message.

Effectively, the bundle-reading code lied to the rest of the compiler by hiding
the placeholders in these messages so they wouldn't confuse the ReplaceMessages
pass.

I'm about to add a new message definition method for ICU-format messages, which
will allow the compiler to actually recognize ICU message placeholders. For
these new messages the compiler WILL recognize the placeholders in the JS code,
and the ReplaceMessages pass will be confused if those placeholders don't
appear in the translated messages read from the translation bundle.

So, the logic for deciding how to handle these placeholders needs to move
from the bundle-reading code to the ReplaceMessages pass.

PiperOrigin-RevId: 475371809