Switch the "most interesting cases" to jspecify @Nullable
Since the jspecify is a type use annotation the rules are sometimes different
about exactly where it can go, and this change moves those affected annotations
to the allowed places.

In these cases, the annotation comes after the first part of the name, which is
awkward, but in these particular cases I still didn't think it was reasonable
to import the short name.

PiperOrigin-RevId: 475963098