Fix template inference bug on repeated `@template T` definitions in type hierarchy
The problem: a TemplateTypeMap for a subclass contains, in order, all the keys in the superclass's TemplateTypeMap + all the subclass keys.

When resolving a type annotation like `/** @type {T} */` in a subclass's scope, the typechecker previously searched for the first template named 'T'. This would resolve to the superclass 'T', not the subclass 'T'. This CL reverses that search so we visit all subclass template types first before looking at the superclass.

This seems slightly brittle: an alternative might be modeling TemplateTypeMaps similar to PropertyMaps, where we distinguish between 'own template types' and 'inherited template types'. But that would be a more complicated change and possibly more expensive in terms of memory - we only need one PropertyMap per class, but we need N TemplateTypeMaps per N instances of Array<string>,Array<number>,etc.

PiperOrigin-RevId: 476136419