Store message placeholder names in original format
When creating a JsMessage object's placeholders,
store their names in the format that matches the source from which they were taken,
instead of always coercing them to lowerCamelCase.

Use the stored format of the name when matching the placeholder with its
example and original_code values.

When I implement ICU message templates, the placeholder names in both the
message string and the object literals will be in UPPER_SNAKE_CASE, and we need
those to match each other without having to reformat either one.

PiperOrigin-RevId: 478243816