Avoid failure due to invalid AST shape in `ConvertToTypedInterface`
Don't use `NodeUtil.removeChild()` in the `visit()` method of `ConvertToTypedInterface`,
becaues the AST is likely to be in an invalid shape at that point.
Instead just do `n.detach()`.

`NodeUtil.removeChild()` depends on the AST having a valid shape and throws an exception
if it notices it is not valid.

PiperOrigin-RevId: 480919751