Remove rarely-used JSError.make overrides taking a CheckLevel
Almost all of the time, there's no reason to pass a CheckLevel to JSError.make because a DiagnosticType already has a default level. A few of the existing uses of this method can be easily refactored to not pass a CheckLevel.

This functionality is still available via the Builder API to support the few existing use sites that actually do want to override the CheckLevel of a DiagnosticType: StrictModeCheck and RhinoErrorReporter.

PiperOrigin-RevId: 482595743