Add explicit support for `goog.i18n.messages.declareIcuTemplate()`
This method is intended to allow specification of ICU-formatted messages
including specification of example text and / or original code samples
for placeholders in them.

It was already possible to pass an ICU-formatted message to
`goog.getMsg()`, but the generated XMB entries would always treat the
message as a single string with no placeholder elements.

The actual definition of this new method has not yet been added to
closure-library, so this isn't usable yet.

Also, the method for identifying these calls is brittle,
so I'm considering how I could make it more reliable.

PiperOrigin-RevId: 488766289