Fix 2 for object literal type inference
The first attempt in cl/497979576 broke some targets because properties
were dropped from object literals that were expected in other parts of
the code. This fix makes sure that all properties before the spread
operator are added. Note that any properties after the spread operator
are still dropped.

Description of the first fix:

Type inference may visit the same object literal multiple times. If it contains a spread operator, the type of the literal is converted to the global type Object. A subsequent pass over the same literal could later add inferred properties to type Object, which is wrong.

PiperOrigin-RevId: 498473432