Stop serializing inline source maps (base64-encoded "data urls") in TypedAST
There are 2 kinds of source maps.
#### 1) Input source map
A source map for a file passed as an input to JSCompiler. This can be passed in one of two ways:
1) the `--source_map_input_files` flag
2) a special comment embedded into the input JS file: `//# sourceMappingURL=<url>`
Where the "<url>" points to the source map content.

#### 2) Inline source map
An input source map passed by embedding a sourceMappingURL as a base64-encoded "data url".

We want TypedAST to support input source maps and not inline source maps. This change stops serializing inline source map URLs in TypedAST.

PiperOrigin-RevId: 500744410