Do not support sourcemappingURLs with a configured directory for TypedAST builds.
Previously, we were serializing the **path** of the sourcemap file on disk as the `sourceMappingURL` in lazyAst.

With this change, we will serialize the **name** of the sourcemap file.

This only affects TypedAST builds.

E.g: We will now serialize `a.js.map` instead of `foo/bar/a.js.map`.

PiperOrigin-RevId: 504935372