Precalculate num bytes and numLines when reading source files.
These are often used for recording metrics and calculating them is pretty trivial so doing it slightly more eagerly is reasonable.

lineOffsets is _not_ precalculated since it is only used for warning formatting which should be rare.

These values will persist even after a call to `clearCachedSource()` since they are very cheap to store, this does introduce the possibility that they hold 'stale' values if the source file changes.  Docs have been updated to address this.

PiperOrigin-RevId: 507485430