for-of on a non-iterable, non-arraylike should throw an error.
The utility JS code used when transpiling a for-of loop down to ES5
checks for the `Symbol.iterator` property first, then falls back to
iterating as if the thing to be iterated is an array-like-thing.
This can end up hiding a `TypedError` that should have been thrown.

Check to make sure the thing to iterate over has a `length` property
before just treating it like an array when transpiling a for-of loop.
If it doesn't, throw a `TypeError`.

PiperOrigin-RevId: 509969328