Gzip save files for multistage compilation
The save files scale linearly with the size of the application and can thus be quite large. By gzipping we can reduce sizes substantially, which should save disk space.  Gzip is fairly fast and given that we can expect a reduction in file-io as a result this may or may not be a wall time regression.

PiperOrigin-RevId: 511888713