Fix a bug in CoalesceVariableNames pass where we generate incorrect code with `usePseudoNames=true`
When generating pseudo names, the CoalesceVariableNames pass tries to find all names that are coalesced together with a given coalescedVar (i.e. their liveness does not overlap and they are on the same color-partition). To do that, it iterates over all the variables stored in the liveness object and calls `getPartitionSuperNode()` on them. This method is “order-sensitive” (i.e. the first node to retrieve its partition becomes the super node of that partition). This creates a problem when >1 variables exist in a partition - we get a different supernode with/without usePseudoNames option. This influences which var we “redeclare” (the “supernode” var) and which var’s decl we “remove” (other nodes in that partition).

In this CL, I am adding a method to query the supernode relationship without modifying the `colorToNodeMap`. I use that new method instead of `getPartitionSuperNode()` to recognize vars that belong to the same partition.

PiperOrigin-RevId: 518297880