Serialize function literal colors as the top object type, instead of individually
Only functions that 1) have an instance type, i.e. are an interface or constructor or 2) have a @closurePrimitive annotation are serialized separately. All other function colors are essentially identical for optimization purposes - the colors don't track parameter/return types, for example.

This speeds up deserialization slightly and improves memory usage.

NOTE: this also improves property disambiguation in some edge cases. We no longer invalidate /all/ 'declared properties' on function literal types, only properties actually referenced on said function in the AST.

This could cause optimization issues in code that is poorly typed. The only problem I saw in practice was in aggregateerror.js - it failed to declare the relationship between the AggregateError extern and the polyfill.

PiperOrigin-RevId: 522653019