`CompilerTestCase#parseExpectedJs()` - transpile immediately before normalize
In `CompilerTestCase#parseExpectedJs()` move the transpile passes to
immediately before the normalize pass.

This more accurately reflects real runtime behavior and is also necessary,
because we're about to start moving transpilation passes after normalization
and make at least some of them depend on it.

It will also be necessary to do a similar change in the `testInternal` method.

PiperOrigin-RevId: 530673016