CompilerTestCase - move transpilation immediately before normalization
This change makes CompilerTestCase behavior better match what actually
happens in normal execution.

It is also necessary, because I'm about to start moving transpilation
passes after normalization and make some of them depend on it.

NOTE: This change also fixes what appears to have been
a long-standing bug in the deserialization method used for testing.
It failed to correctly add externs inputs.

The change in pass ordering exposed this bug.

PiperOrigin-RevId: 530722021