Make variable renaming by `Normalize` optional
I'm about to make some transpilation passes depend on and maintain
normalization, so they can be simplified and eventually be moved after some
optimizations.

For some transpile-only use cases that involve Angular,
renaming parameter names will break runtime parameter injection.

The transpilation passes won't depend on forcing all variable names to be
unique.
That is only needed for optimizations.

So, this change adds an option to `Normalize` to disable the variable
renaming it does.
This is intended to be used only for transpile-only cases.

PiperOrigin-RevId: 531098280