Stop trying to include `'use strict';` in CommonJS module output
The `Es6RewriteModulesToCommonJsModules` pass generates a function which it was trying
to have start with a `'use strict';` pragma statement.

However, this is ineffective in nearly every case because optimizations and transpilations
are likely to either move statements ahead of the pragma or remove it entirely.

Trying to mark the function body for strict mode execution in the correct way is also
ineffective as described in the comments I've added to the code.

This problem was revealed by work that is aiming at moving transpilation passes
after normalization.

Normalization will move local function declarations to the beginning of the function body.

PiperOrigin-RevId: 531549428