Fix transpilation of a let/const in a loop, used in an object getter/setter
When seeing references to 'let/const' loop vars in functions, transpilation currently wraps the function in an IIFE, to capture the 'let/const's values. This IIFE wrapping is not syntactically possible for getter/setter defs. Instead this change wraps the entire containing object literal in an IIFE.

Note: class getters/setters and methods are always transpiled away before this pass runs and are not affected.

Fixes https://github.com/google/closure-compiler/issues/3599.

PiperOrigin-RevId: 532132957