Include IS_CONSTANT_NAME property on newly created name nodes
I'm about to move a usage of AstFactory.createName() to occur after normalization.
It turns out that one thing normalization enforces is that all references to a
variable flagged as constant must have the `IS_CONSTANT_NAME` property with a
`true` value.
So, `AstFactory` needs to enforce this when it can.

PiperOrigin-RevId: 534182011