enableNormalizeExpectedOutput() in individual test cases
For a few test classes stop normalizing the expected output for all test cases and instead do it
only for the few that currently need that behavior.

Ideally most, and possibly all, of these test cases should be further modified to explicitly
state the actual expected output instead of a non-normalized form. That would be less confusing
when test failures occur.

PiperOrigin-RevId: 537908852