Refactor ReplaceCssNames
The intentions of this refactor are:

1. Expose the `goog.getCssName()` calls found in the AST as a list,
   instead of modifying them as they are found.
2. Use a coding pattern that will be easier to modify when we discover
   future changes that need to be made.
3. Improve readability.

The primary motivation behind this refactor is to prepare the addition of
support for a new JS coding pattern that will require some `goog.getCssName()`
calls be handled a bit differently from others.

This refactor also introduces a new error message.

The code pattern `goog.getCssName(goog.getCssName('n1'), 'n2')` was
previously reported as an error with a message that claimed that the inner
call to `goog.getCssName()` was a string literal, even printing the converted
value of that call. This change introduces a separate error message explicitly
for that case.

PiperOrigin-RevId: 539797260