Don't crash when constructing a string key with an object pattern on the RHS
Currently, code such as

```
let {["x"]:{["y"]:bar}}={["x"]:{["y"]:"foo"}}
```

will cause the compiler to throw an IllegalStateException while converting the
computed properties on the LHS to dotted properties. When converting

```
{["y"]:bar}
```

to dotted properties, the compiler asserts that the RHS of the new string key
is an expression. In this case, the RHS is an object pattern, which the
compiler does not consider to be an expression, so an assertion error is
thrown.

This change loosens the check so that object patterns are permitted on the RHS
of a string key.

PiperOrigin-RevId: 540110037