Drop extra labeled blocks from loop let/const transpilation
tl;dr - Stop generating extra labeled blocks in loops containing closures when
transpiling `let` and `const`.

This is a code size and simplicity win.

Details:

The type inference pass understands `let` and `const` variable scoping very
well, but it does a poor job handling the loop state objects that we must
generate in their place when transforming all `let` and `const` variables into
`var`s. For this reason, we have been adding extra labeled blocks to those
loops in order to allow us to put the loop object update assignment at the
bottom of the loop. This kept the order of assignments and usages in an order
that type inference could handle better.

Now that type checking always happens before transpilation of `let` and
`const`, it is no longer necessary to add extra labeled blocks to loops
containing closures.

PiperOrigin-RevId: 546902020