Delete unused pass CheckMissingOverrideTypes and related tooling
Context: It was added to aid TS migration efforts using the global_types_fixing_tool. But the pass was not useful as adding the missing JSDoc types tightened the existing types of `ts_library` targets, producing downstream type errors. The pass was kept disabled since and is unlikely to be reused for migration anytime soon.
PiperOrigin-RevId: 547026935