`LoopClosureTransformer`: Move logic from visit() to shouldTraverse()
This is the first step toward improving the behavior and readability of
`Es6RewriteBlockScopedDeclaration.LoopClosureTransformer`.
It is has been doing work in `visit()` that could easily be done in
`shouldTraverse()` instead. Doing it there will allow later changes
to save that information in a more convenient form as it is discovered.

`LoopClosureTransformer` is a `NodeTraversal.Callback` class that handles
converting `let`/`const` variables declared in loops and used in closures
declared in those loops.

It is very inefficient and hard to read because it fails to gather
information when it is conveniently available on the way down the AST.

PiperOrigin-RevId: 547480793