Delete incorrect removal of FREE_CALL property
Es6RewriteBlockScopedDeclaration has been incorrectly rewriting calls to
loop-local let/const variables since ~2015. It's a fairly harmless mistake,
but we should fix it.

To handle closures correctly loop-local variables have to be transformed
into properties on an object that gets replaced on each iteration of the loop
with a new object.

This means that `localVar()` will get rewritten as
`$jscomp$loop$var$0.localVarAsProp()`.

Except, that's wrong, actually, because if it gets rewritten like that,
then the `$jscomp$loop$var$0` value will get passed into the function as
the `this` value. So the way the call is executed has now changed.

It should really look like `(0,$jscomp$loop$var$0.localVarAsProp)()`,
which is what you would get if you simply replaced the variable name
with the property reference, but there's a snippet of code that was
added in 2015 as part of a bugfix that explicitly removes the
FREE_CALL property from the CALL node.

I would guess that the `(0,...)()` test case output looked weird to the person
doing the bugfix and they "fixed" that problem by removing the FREE_CALL
property.

PiperOrigin-RevId: 547900205