Clean up changeDeclarationsToAssignments()
This method didn't just change declarations.
It changed all loop-local variable references used in closures.
Submethods it called also had misleading names and the logic
was messy because a single method tried to handle both the
declaration of a variable and references to a variable.

This change refactors and renames the method and its callees.

PiperOrigin-RevId: 548198756