Split Es6RenameReferences into 2 classes
When this class is used as part of ES6 module transpliation,
it only needs to modify JSDoc.

When it is used in every other location it is now running after type checking,
so JSDoc modification to rename types is no longer necessary.

This code can be significantly simplified and some performance improvements
made by splitting this into 2 classes. One only renames JSDoc type references,
the other only renames variables in the JS code.

PiperOrigin-RevId: 548858650