Allow i.js to handle public fields.
The i.js file supports the syntax for public fields, allows non-static and static fields to be declared, and the assignment portion for public fields is stripped off. If the public field is a `COMPUTED_FIELD_DEF`, we will remove it from i.js. If a public field is already declared and updates are made to the field, then the updates will be removed from i.js. The unit tests reflect the behavior for supporting public fields.

PiperOrigin-RevId: 549777954