Added public field `this` support to ClosureCheckModule for false positives.
Before, if a public field was assigned using the `this.property` format, ClosureCheckModule would throw `JSC_GOOG_MODULE_REFERENCES_THIS`. ClosureCheckModule now checks if the assignment is allowed given that the `this` is not inside a Goog module body which is done by updating the hoist scope. The unit tests demonstrate this new support feature for public fields.

PiperOrigin-RevId: 551033593