Implemented detached check to i.js for public fields.
There are test cases where we do not want to call `simplify()` for public fields given that they are detached. When we are in `processDeclaration()`, we will now check if the declaration is detached and break out of the method so we do not call `simplify`. Unit tests have been added to show public field support.

PiperOrigin-RevId: 551272612