Invert Normalization variable renaming for transpile-only cases
Instead of skipping variable renaming for the transpile-only case, we should be
doing the same rename-inversion that is done for full builds when variable
renaming is supposed to be disabled.

This will allow the transpilation passes that run after normalization to fully
rely on all variables being unique.

The rename inversion maintains the user's "no renaming" expectations and avoids
breaking runtime behavior that depends on function parameter names. In
particular Angular likes to use function parameter names to control injection
of parameter values at runtime.

Bonus: Fix the CompilerTestCase externs diff. It was backwards.
PiperOrigin-RevId: 551422263