Move Es6RewriteClassExtends and Es6ExtractClasses passes above RewriteClassMembers in the transpilation order.
This change allows RewriteClassMembers to correctly transpile public fields when the class is located in a location that doesn't easily allow multiple statements such as declaring a class inside of a new call (e.g. `new (class Foo { static x; })()`).

For that to work Es6ExtractClasses was also changed to transpile all classes
instead of only transpiling when it detects that Es6 classes are not supported for the output language.

PiperOrigin-RevId: 552519883