Allow the `super` keyword calling a function to be used to instantiate public fields.
Before, when assigning a public field through the `super` keyword calling a function, we would throw `JSC_INVALID_SUPER_ACCESS`. We now check to see if we are in the context of a `MEMBER_FIELD_DEF` to allow the `super` keyword to call a function. A new class called `SuperPropertyAccessAllowedContext`, is now used to create a new context for static blocks and fields given their same behavior.

PiperOrigin-RevId: 552580668