Print config more reliably
The config printing was only happening when compile() or compileModules() was
called or when executed by AbstractCommandLineRunner::doRun(), but that is
sometimes redundant, and sometimes doesn't happen at all.

I've moved this action to the end of the `initModules()` method, which must
always be executed right before performing a compilation action.

This CL also disables debug logging for CompilerTestCase tests by default.
Without that part of the change I was having test timeouts for some tests.
I believe that was because those tests started wasting time printing out
configuration information to debug log files.

PiperOrigin-RevId: 552964960