Allow public fields with the `@final` JsDoc to be assigned once inside a constructor.
Originally, the `CheckAccessControls` pass did not support field assignment inside of a constructor and would throw the `JSC_FINAL_PROPERTY_OVERRIDEN` error. We now allow unassigned public fields with the `@final` JsDoc to be assigned once in the constructor. Allowing public fields to be recognized puts the team closer to flipping the flag to turn on public fields for google3.

PiperOrigin-RevId: 553601386