Update the ASTValidator to expect MODULE_BODY statements anywhere in the SCRIPT (not just at the top) and validate the module contents.
From http://b/294420383, we know that modules can exist during the transpiler passes and the compiler inserts some code above the goog.modules statements.
However, ASTValidator crashes if it notices that a MODULE_BODY node is not the first child of a SCRIPT.

This CL updates the validator to expect MODULE_BODY statement and validates the module contents.

PiperOrigin-RevId: 555635449