Stop inserting code into a module inside `getNodeForCodeInsertion`
The method is intended to be used for finding an code insertion point within a given JSChunk. This is the first CompilerInput in that chunk, or the first CompilerInput if no chunk is provided.

The typical requirement for inserting code into another CompilerInput is that the code be globally visible (i.e. within a SCRIPT). A check already exists to ensure that the first CompilerInput within the given chunk is a SCRIPT. This CL simply extends that check to ensure that the first CompilerInput is a SCRIPT even if no chunk is provided.

PiperOrigin-RevId: 559959778